+++ global2.h
@@ -54,7 +54,8 @@ namespace PH
 			void global_clean();
 
 			//--------------------------------------------------------------
-			int insert_op(uint64_t key,uint64_t value_size,unsigned char* value);
+//			int insert_op(uint64_t key,uint64_t value_size,unsigned char* value);
+			int insert_op(uint64_t key,int value_size,unsigned char* value);
 			int read_op(uint64_t key,unsigned char* buf);
 			int read_op(uint64_t key,std::string *value);
 			int delete_op(uint64_t key);
